release 2.1.0: initial PyPI release 🎉  (#6)
This PR does not add any new features to the package but brings multiple
nice infra improvements:
- the project now uses [uv](https://docs.astral.sh/uv/) as package
manager. This greatly improves the installation times and streamlines
multiple commands to install, build, test and upload the wheels to PyPI.
- we also uploaded the first PyPI wheels 🎉 
- the CI workflows have been reworked to offer test releases triggered
by a GitHub label and uploaded to [TestPyPI](test.pypi.org) and actual
releases.